PALEOmodel

PALEOmodel.Run — Type

Run

Container for model and output.

Fields

model::Union{Nothing, PB.Model}: The model instance.

output::Union{Nothing, AbstractOutputWriter}: model output

Create and initialize

PALEOmodel.initialize! — Function

initialize!(model::PB.Model; kwargs...) -> (initial_state, modeldata)

Initialize model and return initial_state Vector and modeldata struct

Keywords:

eltype::Type=Float64: default data type to use for model arrays

eltypemap=Dict{String, DataType}: Dict of data types to look up Variable :datatype attribute

pickup_output=nothing: OutputWriter with pickup data to initialise from

threadsafe=false: true to create thread safe Atomic Variables where :atomic attribute = true

method_barrier=nothing: thread barrier to add to dispatch lists if threadsafe==true

expect_hostdep_varnames=["global.tforce"]: non-state-Variable host-dependent Variable names expected

Integrate

DifferentialEquations solvers

PALEOmodel.ODE.integrate — Function

integrate(run, initial_state, modeldata, tspan [; kwargs...] )

Integrate run.model as an ODE or as a DAE with constant mass matrix, and write to outputwriter

Arguments

run::Run: struct with model::PB.Model to integrate and output field

initial_state::AbstractVector: initial state vector

modeldata::Modeldata: ModelData struct with appropriate element type for forward model

tspan: (tstart, tstop) integration start and stop times

Keywords

alg=Sundials.CVODE_BDF(): ODE algorithm to use

outputwriter=run.output: PALEOmodel.AbstractOutputWriter instance to hold output

solvekwargs=NamedTuple(): NamedTuple of keyword arguments passed through to DifferentialEquations.jl solve (eg to set abstol, reltol, saveat, see https://diffeq.sciml.ai/dev/basics/common_solver_opts/)

jac_ad=:NoJacobian: Jacobian to generate and use (:NoJacobian, :ForwardDiffSparse, :ForwardDiff)

jac_ad_t_sparsity=tspan[1]: model time at which to calculate Jacobian sparsity pattern

steadystate=false: true to use DifferentialEquations.jl SteadyStateProblem (not recommended, see PALEOmodel.SteadyState.steadystate).

BLAS_num_threads=1: number of LinearAlgebra.BLAS threads to use

init_logger=Logging.NullLogger(): default value omits logging from (re)initialization to generate Jacobian modeldata, Logging.CurrentLogger() to include

PALEOmodel.ODE.integrateDAE — Function

integrateDAE(run, initial_state, modeldata, tspan;alg=IDA())

Integrate run.model as a DAE and copy output to outputwriter. Arguments as integrate.

PALEOmodel.ODE.integrateForwardDiff — Function

integrate with argument defaults to use ForwardDiff AD Jacobian

PALEOmodel.ODE.integrateDAEForwardDiff — Function

integrateDAE with argument defaults to use ForwardDiff AD Jacobian

Fixed timestep solvers

PALEOmodel.ODEfixed.integrateEuler — Function

integrateEuler(run, initial_state, modeldata, tspan, Δt; [,outputwriter])

Integrate run.model using first-order Euler with fixed timestep.

Arguments

run::Run: struct with model::PB.Model to integrate and output field

initial_state::AbstractVector: initial state vector

modeldata::Modeldata: ModelData struct with appropriate element type for forward model

tspan: (tstart, toutput1, toutput2, ..., tstop) integration start, output, stop times

Δt: (yr) fixed timestep

[outputwriter: PALEOmodel.AbstractOutputWriter instance to write model output to]

[report_interval: number of timesteps between progress update to console]

PALEOmodel.ODEfixed.integrateSplitEuler — Function

integrateSplitEuler(run, initial_state, modeldata, tspan, Δt_outer, n_inner;
                        cellranges_outer, cellranges_inner,
                        [,outputwriter])

Integrate run.model representing:

\[\frac{dS}{dt} = f_{outer}(t, S) + f_{inner}(t, S)\]

using split first-order Euler with Δt_outer for f_outer and a shorter timestep Δt_outer/n_inner for f_inner.

f_outer is defined by calling PALEOboxes.do_deriv with cellranges_outer hence corresponds to those Reactions with operatorID of cellranges_outer. f_inner is defined by calling PALEOboxes.do_deriv with cellranges_inner hence corresponds to those Reactions with operatorID of cellranges_inner.

NB: the combined time derivative is written to outputwriter.

Arguments

run::Run: struct with model::PB.Model to integrate and output field

initial_state::AbstractVector: initial state vector

modeldata::Modeldata: ModelData struct with appropriate element type for forward model

tspan: (tstart, toutput1, toutput2, ..., tstop) integration start, output, stop times

Δt_outer: (yr) fixed outer timestep

n_inner: number of inner timesteps per outer timestep

Keywords

cellranges_outer: Vector of CellRange with operatorID defining f_outer.

cellranges_inner: Vector of CellRange with operatorID defining f_inner.

[outputwriter: PALEOmodel.AbstractOutputWriter instance to write model output to]

[report_interval: number of outer timesteps between progress update to console]

PALEOmodel.ODEfixed.integrateEulerthreads — Function

integrateEulerthreads(run, initial_state, modeldata, cellranges, tspan, Δt;
    outputwriter=run.output, report_interval=1000)

Integrate run.model using first-order Euler with fixed timestep Δt, with tiling over multiple threads.

Arguments

run::Run: struct with model::PB.Model to integrate and output field

initial_state::AbstractVector: initial state vector

modeldata::Modeldata: ModelData struct with appropriate element type for forward model

cellranges::Vector{Vector{AbstractCellRange}}: Vector of Vector-of-cellranges, one per thread (so length(cellranges) == Threads.nthreads).

tspan: (tstart, toutput1, toutput2, ..., tstop) integration start, output, stop times

Δt: (yr) fixed outer timestep

PALEOmodel.ODEfixed.integrateSplitEulerthreads — Function

integrateSplitEulerthreads(run, initial_state, modeldata, tspan, Δt_outer, n_inner::Int; 
                            cellranges_outer, cellranges_inner, [,outputwriter] [, report_interval])

Integrate run.model using split first-order Euler with Δt_outer for f_outer and a shorter timestep Δt_outer/n_inner for f_inner.

f_outer is defined by calling PALEOboxes.do_deriv with cellrange_outer hence corresponds to those Reactions with operatorID of cellrange_outer. f_inner is defined by calling PALEOboxes.do_deriv with cellrange_inner hence corresponds to those Reactions with operatorID of cellrange_inner.

Uses Threads.nthreads threads and tiling described by cellranges_inner and cellranges_outer (each a Vector of Vector{AbstractCellRange}, one per thread).

Arguments

run::Run: struct with model::PB.Model to integrate and output field

initial_state::AbstractVector: initial state vector

modeldata::Modeldata: ModelData struct with appropriate element type for forward model

tspan: (tstart, toutput1, toutput2, ..., tstop) integration start, output, stop times

Δt_outer: (yr) fixed outer timestep

n_inner: number of inner timesteps per outer timestep (0 for non-split solver)

cellranges_outer::Vector{Vector{AbstractCellRange}}: Vector of list-of-cellranges, one list per thread (so length(cellranges) == Threads.nthreads), with operatorID defining f_outer.

cellranges_inner::Vector{Vector{AbstractCellRange}}: As cellranges_outer, with operatorID defining f_inner.

PALEOmodel.ODELocalIMEX.integrateLocalIMEXEuler — Function

integrateLocalIMEXEuler(run, initial_state, modeldata, tspan, Δt_outer [; kwargs...])

Integrate run.model representing:

\[\frac{dS}{dt} = f_{outer}(t, S) + f_{inner}(t, S)\]

using first-order Euler with Δt_outer for f_outer and implicit first-order Euler for f_inner, where f_inner is local (within-cell, ie no transport), for a single Domain, and uses only StateExplicit and Deriv variables.

f_outer is defined by calling PALEOboxes.do_deriv with cellranges_outer hence corresponds to those Reactions with operatorID of cellranges_outer. f_inner is defined by calling PALEOboxes.do_deriv with cellrange_inner hence corresponds to those Reactions with operatorID of cellrange_inner.

NB: the combined time derivative is written to outputwriter.

Arguments

run::Run: struct with model::PB.Model to integrate and output field

initial_state::AbstractVector: initial state vector

modeldata::Modeldata: ModelData struct with appropriate element type for forward model

tspan: (tstart, toutput1, toutput2, ..., tstop) integration start, output, stop times

Δt_outer: (yr) fixed timestep

Keywords

cellranges_outer: Vector of CellRange with operatorID defining f_outer.

cellrange_inner: A single CellRange with operatorID defining f_inner.

exclude_var_nameroots: State variables that are modified by Reactions in cellrange_inner, but not needed to find implicit solution (ie reaction rates etc don't depend on them).

[outputwriter=run.output: PALEOmodel.AbstractOutputWriter instance to write model output to]

[report_interval=1000: number of outer timesteps between progress update to console]

[Lnorm_inf_max=1e-3: normalized error tolerance for implicit solution]

[niter_max=10]: maximum number of Newton iterations

[request_adchunksize=4]: request ForwardDiff AD chunk size (will be restricted to an upper limit)

Fixed timestep wrappers

PALEOmodel.ODEfixed.integrateFixed — Function

integrateFixed(run, initial_state, modeldata, tspan, Δt_outer;
            timesteppers, outputwriter=run.output, report_interval=1000)

Fixed timestep integration, with time step implemented by timesteppers,

`timesteppers = [ [(timestep_function, cellranges, timestep_ctxt), ...], ...]`

Where timestep_function(model, modeldata, cellranges, timestep_ctxt, touter, Δt, barrier)

PALEOmodel.ODEfixed.integrateFixedthreads — Function

integrateFixedthreads(run, initial_state, modeldata, tspan, Δt_outer;
            timesteppers, outputwriter=run.output, report_interval=1000)

Fixed timestep integration using Threads.nthreads() threads, with time step implemented by timesteppers,

`timesteppers = [ [(timestep_function, cellranges, timestep_ctxt), ... (1 per thread)], ...]`

Where timestep_function(model, modeldata, cellranges, timestep_ctxt, touter, Δt, barrier).

Steady-state solvers (Julia NLsolve based)

PALEOmodel.SteadyState.steadystate — Function

steadystate(run, initial_state, modeldata, tss [; kwargs...] )

Find steady-state solution (using NLsolve.jl package) and write to outputwriter (two records are written, for initial_state and the steady-state solution).

Arguments

run::Run: struct with model::PB.Model to integrate and output field

initial_state::AbstractVector: initial state vector

modeldata::Modeldata: ModelData struct with appropriate element type for forward model

tss: (yr) model tforce time for steady state solution

Optional Keywords

outputwriter=run.output: PALEOmodel.AbstractOutputWriter instance to hold output

initial_time=-1.0: tmodel to write for first output record

solvekwargs=NamedTuple(): NamedTuple of keyword arguments passed through to NLsolve.jl (eg to set method, ftol, iteration, show_trace, store_trace).

jac_ad: :NoJacobian, :ForwardDiffSparse, :ForwardDiff

use_norm=false: true to normalize state variables using PALEO norm_value

BLAS_num_threads=1: number of LinearAlgebra.BLAS threads to use

PALEOmodel.SteadyState.steadystateForwardDiff — Function

steadystate with argument defaults to use ForwardDiff AD Jacobian

PALEOmodel.SteadyState.steadystate_ptc — Function

steadystate_ptc(run, initial_state, modeldata, tss, deltat_initial, tss_max [; kwargs...])

Find steady-state solution and write to outputwriter, using naive pseudo-transient-continuation with first order implicit Euler pseudo-timesteps from tss to tss_max and NLsolve.jl as the non-linear solver.

Each pseudo-timestep solves the nonlinear system S(t+Δt) = S(t) + Δt dS/dt(t+Δt) for S(t+Δt), using a variant of Newton's method.

Initial pseudo-timestep Δt is deltat_initial, this is multiplied by deltat_fac for the next iteration until pseudo-time tss_max is reached. If an iteration fails, Δt is divided by deltat_fac and the iteration retried.

NB: this is a very naive initial implementation, there is currently no reliable error control to adapt pseudo-timesteps to the rate of convergence, so requires some trial-and-error to set an appropiate deltat_fac for each problem.

Keywords

deltat_fac=2.0: factor to increase pseudo-timestep on success

outputwriter=run.output: output destination

solvekwargs=NamedTuple(): arguments to pass through to NLsolve

jac_ad=:NoJacobian: AD Jacobian to use

request_adchunksize=10: ForwardDiff chunk size to request.

jac_cellranges=modeldata.cellranges_all: CellRanges to use for Jacobian calculation (eg to restrict to an approximate Jacobian)

enforce_noneg=false: fail pseudo-timesteps that generate negative values for state variables.

use_norm=false: true to apply PALEO norm_value to state variables

verbose=false: true to detailed output

BLAS_num_threads=1: restrict threads used by Julia BLAS (likely irrelevant if using sparse Jacobian?)

See steadystate for more details of arguments.

PALEOmodel.SteadyState.steadystate_ptcForwardDiff — Function

steadystate_ptc with argument defaults to use ForwardDiff AD Jacobian

Steady-state solvers (Sundials Kinsol based):

PALEOmodel.SteadyStateKinsol.steadystate_ptc — Function

steadystate_ptc(run, initial_state, modeldata, tss, deltat_initial, tss_max; 
    [,deltat_fac=2.0] [,tss_output] [,outputwriter] [,createkwargs] [,solvekwargs]
    [,jac_cellranges] [, use_directional_ad] [, directional_ad_eltypestomap] [,verbose] [,  BLAS_num_threads] )

Find steady-state solution and write to outputwriter, using naive pseudo-transient-continuation with first order implicit Euler pseudo-timesteps and PALEOmodel.Kinsol as the non-linear solver.

Each pseudo-timestep solves the nonlinear system S(t+Δt) = S(t) + Δt dS/dt(t+Δt) for S(t+Δt), using a variant of Newton's method (preconditioned Newton-Krylov, with the Jacobian as preconditioner)

Initial pseudo-timestep Δt is deltat_initial, this is multiplied by deltat_fac for the next iteration until pseudo-time tss_max is reached. If an iteration fails, Δt is divided by deltat_fac and the iteration retried.

NB: this is a very naive initial implementation, there is currently no reliable error control to adapt pseudo-timesteps to the rate of convergence, so requires some trial-and-error to set an appropiate deltat_fac for each problem.

Solver PALEOmodel.Kinsol options are set by arguments createkwargs (passed through to PALEOmodel.Kinsol.kin_create) and solvekwargs (passed through to PALEOmodel.Kinsol.kin_solve).

Preconditioner (Jacobian) calculation can be modified by jac_cellranges, to specify a operatorIDs so use only a subset of Reactions in order to calculate an approximate Jacobian to use as the preconditioner.

If use_directional_ad is true, the Jacobian-vector product will be calculated using automatic differentiation (instead of the default finite difference approximation). directional_ad_eltypestomap can be used to specify Variable :datatype tags (strings) that should be mapped to the AD directional derivative datatype hence included in the AD directional derivative.

PALEOmodel.Kinsol — Module

Kinsol

Minimal Julia wrapper for the Sundials kinsol nonlinear system solver https://computing.llnl.gov/projects/sundials/kinsol

This closely follows the native C interface, as documented in the Kinsol manual, with conversion to-from native Julia types.

The main user-facing functions are Kinsol.kin_create and Kinsol.kin_solve.

PALEOmodel.Kinsol.kin_create — Function

kin_create(f, y0 [; kwargs...]) -> kin

Create and return a kinsol solver context kin, which can then be passed to kin_solve

Arguments

f: Function of form f(fy::Vector{Float64}, y::Vector{Float64}, userdata)

y0::Vector template Vector of initial values (used only to define problem dimension)

Keywords

userdata: optional user data, passed through to f etc.

linear_solver: linear solver to use (only partially implemented, supports :Dense, :Band, :FGMRES)

psolvefun: optional preconditioner solver function (for :FGMRES)

psetupfun: optional preconditioner setup function

jvfun: optional Jacobian*vector function (for :FGMRES)

PALEOmodel.Kinsol.kin_solve — Function

kin_solve(
    kin, y0::Vector;
    [strategy] [, fnormtol] [, mxiter] [, print_level] [,y_scale] [, f_scale] [, noInitSetup]
) -> (y, kin_stats)

Solve nonlinear system using kinsol solver context kin (created by kin_create) and initial conditions y0. Returns solution y and solver statistics kinstats. kinstats.returnflag indicates success/failure.

Field Array

FieldArray provides a generic array type with named dimensions PALEOboxes.NamedDimension and optional coordinates PALEOboxes.FixedCoord for processing of model output.

PALEOmodel.FieldArray — Type

FieldArray

A generic xarray-like or IRIS-like Array with named dimensions and optional coordinates.

NB: this aims to be simple and generic, not efficient !!! Intended for representing model output, not for numerically-intensive calculations.

PALEOmodel.get_array — Function

get_array(obj, ...) -> FieldArray

Get FieldArray from PALEO object obj

PALEOmodel.FieldRecord — Type

FieldRecord{D <: AbstractData, S <: AbstractSpace, ...}

A series of records each containing a PALEOboxes.Field.

Implementation

Fields with array values are stored in records as a Vector of arrays. Fields with single values (fieldsingleelement true) are stored as a Vector of eltype(Field.values).

Output

PALEOmodel.OutputWriters — Module

OutputWriters

Data structures and methods to hold and manage model output.

PALEOmodel.OutputWriters.OutputMemory — Type

OutputMemory

In-memory model output, organized by model Domains.

Field domains::Dict{String, OutputMemoryDomain} contains per-Domain model output.

PALEOmodel.OutputWriters.OutputMemoryDomain — Type

OutputMemoryDomain

In-memory model output, for one model Domain.

Includes an additional coords_record (usually :tmodel, when storing output vs time).

Implementation

data::DataFrame contains columns of same type as FieldRecord.records for each Variable.

PALEOmodel.OutputWriters.save_jld2 — Function

save_jld2(output::OutputMemory, filename)

Save to filename in JLD2 format (NB: filename must either have no extension or have extension .jld2)

PALEOmodel.OutputWriters.load_jld2! — Function

load_jld2!(output::OutputMemory, filename)

Load from filename in JLD2 format, replacing any existing content in output. (NB: filename must either have no extension or have extension .jld2).

Example

julia> output = PALEOmodel.OutputWriters.load_jld2!(PALEOmodel.OutputWriters.OutputMemory(), "savedoutput.jld2")

PALEOmodel.OutputWriters.initialize! — Function

initialize!(output::OutputMemory, model, modeldata, nrecords [;rec_coord=:tmodel])

Initialize from a PALEOboxes::Model, reserving memory for an assumed output dataset of nrecords.

The default for rec_coord is :tmodel, for a sequence of records following the time evolution of the model.

PALEOmodel.OutputWriters.add_record! — Function

add_record!(output::OutputMemory, model, modeldata, rec_coord)

Add an output record for current state of model at record coordinate rec_coord. The usual case (set by initialize!) is that the record coordinate is model time tmodel.

Plot output

RecipesBase.apply_recipe — Method

plot(fa::FieldArray; kwargs...)

Plot recipe that plots fa. If fa has a single dimension, this is suitable for a line-like plot, if two dimensions, a heatmap.

Keywords

swap_xy::Bool=false: true to swap x and y axes

mult_y_coord=1.0: workaround for bugs in Plots.jl heatmap yflip - multiply y coordinate by constant factor.

structfield::Union{Symbol, Nothing}=nothing: use field structfield from a struct-valued array.

map_values=PB.get_total: function to apply to y (for a 1D series) or z (for a 2D heatmap etc) before plotting

labelprefix="": prefix for plot label.

labellist=[]: list of labels to override defaults

labelattribute=nothing: FieldArray attribute to use as label

RecipesBase.apply_recipe — Method

plot(output::AbstractOutputWriter, vars::Union{AbstractString, Vector{<:AbstractString}}, selectargs::NamedTuple=NamedTuple())

Plot recipe that calls PB.get_field(output, var), and passes on to plot(fr::FieldRecord, selectargs)

Additional features:

if vars is a Vector, create a plot series for each element.

if var is of form <domain>.<name>.<structfield>, then set structfield to take a single field from a struct Variable

RecipesBase.apply_recipe — Method

plot(outputs::Vector{<:AbstractOutputWriter}, vars::Union{AbstractString, Vector{<:AbstractString}}, selectargs::NamedTuple=NamedTuple())

Pass through (ie "broadcast") each element output of outputs to plot(output::AbstractOutputWriter, vars, selectargs), adding a labelprefix (index in outputs Vector) to identify each plot series produced.

RecipesBase.apply_recipe — Method

plot(fr::FieldRecord, selectargs::NamedTuple)

Plot recipe that calls get_array(fr; selectargs...) and passes on to plot(fa::FieldArray).

Vector-valued fields in selectargs are broadcasted (generating a separate plot series for each combination)

RecipesBase.apply_recipe — Method

plot(fas::Vector{<:FieldArray}; labelprefix="")

Pass through (ie "broadcast") each element of fas to plot(fa::FieldArray), generating one plot series for each. Adds Vector index to labelprefix.

PALEOmodel.PlotPager — Type

PlotPager(layout, [, kwargs=NamedTuple()])

Accumulate plots into subplots.

layout is supplied to Plots.jl layout keyword, may be an Int or a Tuple (ny, nx), see https://docs.juliaplots.org/latest/

Optional kwargs provides keyword arguments supplied to plot (eg (legendbackgroundcolor=nothing, ) to set all subplot legends to transparent backgrounds)

Usage

julia> pp = PlotPager((2,2))  # 4 panels per screen (2 down, 2 across)
julia> pp(plot(1:3))  # Accumulate
julia> pp(:skip, plot(1:4), plot(1:5), plot(1:6))  # add multiple panels in one command
julia> pp(:newpage) # flush any partial screen and start new page (NB: always add this at end of a sequence!)

Commands

pp(p): accumulate plot p

pp(:skip): leave a blank panel

pp(:newpage): fill with blank panels and start new page

pp(p1, p2, ...): multiple plots/commands in one call

Missing docstring.

Missing docstring for Plot.test_heatmap_edges. Check Documenter's build log for details.

Analyze reaction network

PALEOmodel.ReactionNetwork — Module

ReactionNetwork

Functions to analyize a PALEOboxes.Model that contains a reaction network